+++ hutool-core/src/test/java/cn/hutool/core/io/FileUtilTest.java
@@ -446,7 +446,7 @@ public class FileUtilTest {
 
 		mimeType = FileUtil.getMimeType("test.js");
 		// 在 jdk 11+ 会获取到 text/javascript,而非 自定义的 application/x-javascript
-		List<String> list = ListUtil.of("text/javascript", "application/x-javascript");
+		final List<String> list = ListUtil.of("text/javascript", "application/x-javascript");
 		Assert.assertTrue(list.contains(mimeType));
 
 		// office03
@@ -468,6 +468,10 @@ public class FileUtilTest {
 		// pr#2617@Github
 		mimeType = FileUtil.getMimeType("test.wgt");
 		Assert.assertEquals("application/widget", mimeType);
+
+		// issue#3092
+		mimeType = FileUtil.getMimeType("https://xxx.oss-cn-hangzhou.aliyuncs.com/xxx.webp");
+		Assert.assertEquals("image/webp", mimeType);
 	}
 
 	@Test