+++ hutool-core/src/main/java/cn/hutool/core/thread/SyncFinisher.java
@@ -41,6 +41,10 @@ public class SyncFinisher implements Closeable {
 	 * 结束同步器，用于等待所有worker线程同时结束
 	 */
 	private CountDownLatch endLatch;
+	/**
+	 * 异常处理
+	 */
+	private Thread.UncaughtExceptionHandler exceptionHandler;
 
 	/**
 	 * 构造
@@ -61,6 +65,18 @@ public class SyncFinisher implements Closeable {
 	 */
 	public SyncFinisher setBeginAtSameTime(boolean isBeginAtSameTime) {
 		this.isBeginAtSameTime = isBeginAtSameTime;
+		return this;
+	}
+
+	/**
+	 * 设置异常处理
+	 *
+	 * @param exceptionHandler 异常处理器
+	 * @return this
+	 * @since 5.8.19
+	 */
+	public SyncFinisher setExceptionHandler(final Thread.UncaughtExceptionHandler exceptionHandler) {
+		this.exceptionHandler = exceptionHandler;
 		return this;
 	}
 
@@ -127,10 +143,14 @@ public class SyncFinisher implements Closeable {
 		endLatch = new CountDownLatch(workers.size());
 
 		if (null == this.executorService || this.executorService.isShutdown()) {
-			this.executorService = ThreadUtil.newExecutor(threadSize);
+			this.executorService = buildExecutor();
 		}
 		for (Worker worker : workers) {
-			executorService.submit(worker);
+			if(null != this.exceptionHandler){
+				executorService.execute(worker);
+			} else{
+				executorService.submit(worker);
+			}
 		}
 		// 保证所有worker同时开始
 		this.beginLatch.countDown();
@@ -228,5 +248,17 @@ public class SyncFinisher implements Closeable {
 		 * 任务内容
 		 */
 		public abstract void work();
+	}
+
+	/**
+	 * 构建线程池，加入了自定义的异常处理
+	 *
+	 * @return {@link ExecutorService}
+	 */
+	private ExecutorService buildExecutor() {
+		return ExecutorBuilder.create()
+			.setCorePoolSize(threadSize)
+			.setThreadFactory(new NamedThreadFactory("hutool-", null, false, exceptionHandler))
+			.build();
 	}
 }